Add a default max limit to the logsForRun and eventConnection resolvers (#32028)
## Summary & Motivation

The Dagster UI currently iterates through run logs in chunks, but there are underlying graphql resolvers that let you query for an unbounded number of logs for any number of runs. Apply the same default limit that the UI does to ensure that a single graphql query can’t use too many resources.

## How I Tested These Changes

New test case

## Changelog

The `logsForRun`​ resolvers and `eventConnection`​ resolvers in the Dagster GraphQL API will now apply a default limit of 1000 to the number of logs returned from a single graphql query. The `cursor`​ field in the response can be used to continue iterating through the logs for a given run.