Add "Sign in" and "Create account" buttons
I also reduced the height of the Dagster docs logo by 4px to give us a little more room in the nav, and a fix to crop the logo to the daggy-icon-only in the awkward zone between medium viewport width and mobile. Even in the existing docs prod, there was an awkward viewport width range where the nav items broke out of their flex container and overlapped with the logo. Finally, added text-wrap: nowrap to nav links to prevent "User guide" link from wrapping and blowing out flex wrapper.