[dg] Remove hidden flag from dg scaffold defs inline-component (ENG-113) (#32039)
## Summary & Motivation

Unhide the `inline-component` command in the scaffold defs CLI group.
This command was previously hidden for the initial 1.11 launch as it was
undocumented, but is now ready to be exposed to users.

## How I Tested These Changes

Verified that the `inline-component` command now appears in the CLI help
output when running `dagster-dg scaffold defs --help`.

## Changelog

Added `inline-component` command to the publicly available scaffold
commands in the Dagster CLI.