only include filterFailed and filterMissing checkbox values on reported events if the checkboxes are visible
Summary:
The new flags added in https://github.com/dagster-io/dagster/pull/31878 are checked even if the checkboxes that set them are not visible. Instead, only check them if the checkboxes are visible.

Locally, materialize a partitioned asset, then try to report the materialization.

Fixed an issue where the "Report materialization events" dialog for partitioned assets only worked if the partition was failed or missing.