Ensure that failed partition subset and materialized subset are sourced from the same data in both asset health and partition statuses
Summary:
This prevents an issue where in rare circumstances, asset health and asset partition status could show different results for which partitions had failed, by ensuring that the materialized and failed subsets are always sourced from the same place (Specifically, the AssetMaterializationHealthState object). We still have to fetch the in progress subset separately, since that is not currently available on AssetMaterializationHealthState. This should not result in significant additional data fetching since either the AssetMaterializationHealthState will be cached and quick to compute, or it will fall back to computing its data from the asset status partition cache, the potentially expensive part of which is currently cached at the loader level after https://app.graphite.dev/github/pr/dagster-io/dagster/32108.