[sbc] Use ExitStack for DefsStateStorage context management (#32103)
## Summary & Motivation

Found a pretty nasty bug when attempting to use a StateBackedComponent in a project that used `FivetranWorkspace.load_asset_specs()` (which uses the old StateBackedDefinitionsLoader functionality). The basic issue is that this ends up using a contextmanager that (among other things), constructs an `UnboundInitResourceContext` object.

This object requires an instance to be made available, so it uses the `ephemeral_instance_if_missing()` context manager in the case that an explicit DagsterInstance is not provided, which it is not in this case. This means that we end up `__enter__`ing a separate DagsterInstance object, which in turn sets the current DefsStateStorage to the defs_state_storage of that new ephemeral instance, which happens to be `None` (although it'd also be bad if it was a non-None default value).

There are two core issues here:

- The UnboundInitResourceContext is kinda unholy in how it manages state because it doesn't require you to `__enter__` it, it can't guarantee that `__exit__` is called (and so it also does some cleanup in `__del__`). This logic contains a bug in that it only cleans up the instance if an instance was NOT provided explicitly, but this is actually the opposite of what you would expect. For example, imagine:

```python

with instance_for_test() as instance:
    with UnboundInitResourceContext(instance=instance) as resource_context:
        ...
    # at this point, instance.__exit__ is called because we are
    # exiting the UnboundInitResourceContext and we explicitly
    # provided an instance, even though we haven't left the
    # outer context!
    
    instance.foo()
    instance.bar()
```

In reality, we should only call __exit__ on instances that we created within the UnboundInitResourceContext, as these are the ones that won't get cleaned up through other means.

So the full fix here is to 1) update the __exit__ / __del__ logic 2) use UnboundInitResourceContext as a contextmanager in the appropriate code path


- The second issue is that when a DagsterInstance was `__exit__`ed before, it would always just set the current DefsStateStorage back to `None`, regardless of what the previous value was. Now, we have a special-purpose context manager that handles state transitions properly, so that once we `__exit__` the ephemeral instance, we get our original DefsStateStorage back.


## How I Tested These Changes

## Changelog

NOCHANGELOG