[sbc] ensure DAGSTER_HOME is set for refresh-defs-state (#32121)
## Summary & Motivation

Ran into this while experimenting -- you get confusing behavior if DAGSTER_HOME is not set, because the command will successfully write state, but it will write it to a temporary directory and instantly be cleared away. so if you have a separate terminal running `dg dev`, that will not actually be impacted by that updated state.

## How I Tested These Changes

unit

## Changelog

NOCHANGELOG