[ui] Fix missing loading bar on the RunsFeedTable (#31582)
It looks like the Runs feed table has not been showing it’s
indeterminate loading bar, and Daniel flagged that it was confusing when
switching tabs, since there’s a delay before the rows reflect the new
tab.

This is happening because the IndeterminateLoadingBar is rendered inside
the `overflow:hidden` portion of the run feed table. Both code paths
inside `content()` render the loading bar, so I moved it up one layer so
that it is rendered in the same place but outside the overflow: hidden
container.

Sidenote: The loading bar has a height of 2px even when it's not
animating, which does create a bit of visual space in the design that
isn't there otherwise. Propose we consider making it 2px transparent and
not 2px gray, but I'll post in the design channel.
 
After:

<img width="1754" height="397" alt="image"
src="https://github.com/user-attachments/assets/f6dfc651-cf00-4085-b422-8cb261b8ab4a"
/>

Co-authored-by: bengotow <bgotow@elementl.com>