support inf/-inf in float metadata values (#32138)
Summary:
These are valid python float values but not graphql float values. Add a floatRepr field similar to intRepr so that there is some way to expose those values in the graphql layer.

## How I Tested These Changes
BK

## Changelog
Fixed an issue where setting a FloatMetadataValue to float('inf') or float('-inf') would cause an error when loading that metadata over graphql.