FE-949: Improve UI performance for editing large number of assets in launchpad (#32135)
## Summary & Motivation

I had trouble reproing this at first, but I pasted 30k lines of YAML
into the launchpad and observed that it became sluggish and hard to
edit. I think this user may have assets with a lot of default config,
which means that even if you don’t type a lot there is a large
`rootDefaultYaml` document being parsed on every keystroke.

The “Scaffold default / missing config” options are located in the
bottom drawer, which is already debounced and grays out while it’s
rechecking your config. This PR just moves the calculation of the
`anyDefaultsToExpand` state into that debounced function.

I don’t think this calls for a webworker, it’s too slow to happen every
keystroke but still only ~100ms in my exaggerated test case.

## How I Tested These Changes

After this change, I tried common actions like deleting all the Yaml +
refilling it with the defaults to make sure that the debounce is fast
enough that the button becoming enabled after a slight delay isn't
annoying. Even though this button wasn't debounced before, it was
actually pretty difficult to click without waiting for the debounce,
since the async check grays out the whole bottom panel.

## Changelog

[ui] Typing large YAML documents in the launchpad when default config is
present is now more performant.

Co-authored-by: bengotow <bgotow@elementl.com>