If a backfill fails, cancel all the runs it launched (#31935)
Updates how we handle backfill failure. Instead of marking the backfill as failed and then being done with it, we mark it as FAILING. This works like the CANCELING status. On the following iterations, the daemon will cancel all of the runs launched by the backfill. Once all runs are in a terminal state, the backfill is marked FAILED