[ui] Support code location wildcard on job and timeline input (#32193)
## Summary & Motivation

We have received reports that users would like to be able to use wildcard filtering on code locations when searching in the jobs list and on the timeline page when searching across automations. Add support for this.

## How I Tested These Changes

Try various search strings like `code_location:"foo*"` on timeline page and jobs page, verify that filtering works correctly.

## Changelog

[ui] Allow searching across code locations with `*` wildcard in selection inputs for jobs and automations.