DOC-1413 Update docs for Observe GA (#32122)
## Summary & Motivation

* Moved Observe docs out of Labs into dedicated section:
https://nikki-docs-observe-ga.archive.dagster-docs.io/guides/observe
* Updated Observe docs content to remove "summer update" and related
language
* Turned "Monitor & debug" into "Log & debug" section
* Refactored remaining docs and update content as needed
* Refactored docs snippets and images as needed
* Added `dagster-plus-feature` tag to pages about Dagster+ features.
(**Note:** I used this tag instead of `dagster-plus` because I want to
reserve `dagster-plus` for a time when we want to tag content as
applicable to Dagster+, OSS, or both. In that case, most pages would
have the `dagster-plus` tag, since most content is relevant to Dagster+
users, but the pages may not necessarily be about Dagster+ features.)
* Added redirects

## Remaining questions

- I tried to add `FreshnessPolicy` to API docs, but got an import error,
even after referencing `dagster.preview.freshness.FreshnessPolicy`.
Should `FreshnessPolicy` be moved out of preview/internal?

## How I Tested These Changes

Local build

## Changelog

> Insert changelog entry or delete this section.

---------

Signed-off-by: nikki everett <nikki@dagsterlabs.com>
Co-authored-by: Matt Krukowski <m.krukowski@dagsterlabs.com>