[ui] Upgrade Typescript to 5.9.2 (#32230)
## Summary & Motivation

Bump TS and related dependencies to match the version already running by
default in vscode.

I had to make a couple of changes to eliminate TS errors, most notably
in `generateObjectHash`. I'll note it inline. The related test passes,
so I think this might be TS being a little overzealous.

ui-components uses the public package of our eslint-config, so I'll need
to publish the config, then update the dependency in ui-components. In
the meantime it'll print a warning on lint because the typescript-eslint
tools expect a lower TS version.

## How I Tested These Changes

TS, lint, jest.