[dagster-databricks] Use single multi-asset per Databricks task in DatabricksAssetBundleComponent (#32235)
## Summary & Motivation

Related to #32214 

Previously, all databricks assets were created with a single
multi-asset, which complexified how tasks were submitted and assets
events were emitted.

This PR changes that behavior so that a single Databricks task maps to a
single multi-asset - Changes in #32214 let users define which assets are
included in a Databricks Task. These assets should be created under a
single AssetsDefinition using a multi-asset.

This pattern is similar to the one we are using for Fivetran (one
connector, one multi-asset). Submitting a task materializes all assets
included in this task, so handling them should be done in a single op.

## How I Tested These Changes

Updated tests with BK