Speed up assetsLatestInfo calculation (#32280)
## Summary & Motivation
The old code here was very mystifying to me. We fetch the nodes, make a
map using info in the nodes, then only use the key in that map, and not
the values.

The only benefit I can see is that before it was filering out nodes not
in the graph, but it seems perfectly reasonable to fetch the latest info
for a node not in the graph.

## How I Tested These Changes
BK

## Changelog