[ui] Limit the amount of data needed for EventPopover (#32295)
## Summary & Motivation

When querying for data on the catalog "latest materialization" pez, just query for the bare necessities.

Right now we're querying for way too much and it can be painfully slow even though we don't need all of the data we're asking for.

## How I Tested These Changes

View catalog list, verify that this query is consistently much quicker than what we have in prod. Verify that the event pez render correctly.