Remove unneccesary time window validation checks in asset backfill logic (#32303)
## Summary & Motivation

This was excessively strict - if the partitions def has changed in a safe way that did not remove any partitions, it should be fine to continue the backfill. We actually have tests that verify that case already, they just were not running this method.

## How I Tested These Changes

BK, dry run on a backfill that was previously failing now completes

## Changelog

Fixed an issue where backfills would fail if a TimeWindowPartitionsDefinition's start date was changed in the middle of the backfill, even if it did not remove any of the targeted partitions.