[sbc] Fix context var error (#32267)
## Summary & Motivation

This issue crops up as a side effect of other errors when inside of a webserver that does fancy async things.

The actual error is generally harmless (the process was already in the process of erroring, this just gets added to the stack trace), but we should handle this more gracefully in general and the original behavior here was to set the current storage to None on exit anyway (the contextvar stuff is really just necessary for some pretty niche scenarios) so messing with this isn't the end of the world.

## How I Tested These Changes

verified the test produced the originally observed error, and now no longer does

## Changelog

NOCHANGELOG