[ui] Use execution plan asset keys to determine asset lineage link on Run page (#32260)
## Summary & Motivation

Now that @OwenKephart has added the execution plan `assetKeys` to runs,
we can use this to correctly build an href from a run to view asset
lineage.

- If there is an asset selection, link to that.
- Else, if there is a list of asset keys in the execution plan, link to
that.

## How I Tested These Changes

View a run that has an `assetSelection`, verify that it links correctly.
View a run that only has asset keys on `executionPlan`, verify that it
links correctly.

## Changelog

[ui] Fixed the link to "View asset lineage" on runs that don't specify
an asset selection.