Don't fetch the full parent job when launching subset runs for implicit asset jobs (#31566)
Summary:
This parent job can be very large for ad-hoc asset materializations, and we don't display it anywhere the way that we do for 'regular' subset jobs. Avoid loading the full job (or even the serialized repository object) for these jobs.

## How I Tested These Changes
BK, launch an ad-hoc materailization of an asset and inspect the resulting run object