revert recent changes to codemirror-yaml
Summary:
Still digging but some of the changes to this file in https://github.com/dagster-io/dagster/pull/32255 seem to have stopped auto-complete from firing in some situations. I imagine this change will also make lint fail given the nature of the blamerev so will have to sort that out.