adjust make_graphql_context_test_suite fixtures (#32307)
## Summary & Motivation
I'd like to subclass the graphql context test suites to add more hooks
into the context, while maintaining access to the superclass's graphql
context.

The cleanest way I found to do this was to separate out the construction
of the context and the fixture function. Each subclass can add onto the
implementation of the context construction, and slightly change the
fixture as needed. The overriding of arguments / usage of fixtures
required lots of type ignoring.

## How I Tested These Changes
BK