gql endpoint for fetching the latest materialization failure timestamp (#32329)
Adds an endpoint for fetching the time the asset most recently failed to materialize. We'll use this in the `Failed materializations`​ tab on the homepage

For a perf improvement we can start storing this timestamp in streamline. Fetching via the asset record will unblock the ui sooner, so start there and i'll do the streamline work as a followup https://linear.app/dagster-labs/issue/OPER-2443/source-timestamps-used-in-the-homepage-from-streamline