fix break condition in parentsPoppingItemsDeeperThan preventing auto-complete from working (#32333)
## Summary & Motivation
Fixes an issue introduced in https://github.com/dagster-io/dagster/pull/32255. The old while loop condition was >=, so the break condition should be <, not >=.

## How I Tested These Changes
Launchpad autocomplete for nested config fields now works as expected.

## Changelog
Fixed an issue introduced in the 1.11.12 release where auto-complete in the Launchpad for nested fields stopped working.