fix backfil not moving runs from failing to failed because of unloadable partitions (#32256)
Fixes an issue with the backfill daemon where if a backfill is in `CANCELING`​ or `FAILING`​ status and the asset backfill data becomes unloadable (ex. a partitions def changed) the iteration for the backfill fails immediately, rather than running the cancellation logic for the runs of the backfill. This means that the backfill is permanently stuck in `CANCELED`​ or `FAILING`​.

This PR updates the backfill iteration so that if a backfill is `CANCELED`​ or `FAILING`​ and the backfill data is unloadable, we continue with canceling the runs that were launched by the backfill, but skip updating the asset backfill data. This means that the asset backfill data will be inaccurate for that backfill though, since it won't move the assets in canceled runs from in progress to failed.



### Changelog 

[bugfix] Fixed a bug in the backfill daemon where an asset backfill with `CANCELING`​ or `FAILING`​ status could become permanently stuck in `CANCELING`​ or `FAILING`​ if the partitions definitions of the assets changed.