client module disables slash hotkey
Check if Scout is in focus and then, if so, stop "/" event from opening Algolia window.

I explored opening a PR with docusaurus to add more native support for Algolia keyboardShortcuts but learned docusaurus is still running an Algolia version that doesn't support that param, and so we'll likely need to wait. Because this only blocks events that fire within scout widget, I think it's pretty harmless.