DOC-1392 client module disables slash hotkey (#32340)
## Summary & Motivation


https://linear.app/dagster-labs/issue/DOC-1392/algolia-hot-key-steals-focus-from-ask-ai-input-area

Check if Scout is in focus and then, if so, stop "/" event from opening
Algolia window.

I explored opening a PR with docusaurus to add more native support for
Algolia keyboardShortcuts but learned docusaurus is still running an
Algolia version that doesn't support that param, and so we'll likely
need to wait. Because this only blocks events that fire within scout
widget, I think it's pretty harmless.

## How I Tested These Changes

I modified the docusaurus configuration to load Algolia docsearch widget
locally and confirm everything behaves as expected. I also added a
conditional console log to confirm that "/" events are handled normally
if scout is in focus (but deleted console test before committing).

## Changelog

> Insert changelog entry or delete this section.

---------

Co-authored-by: Kyle P <kylespalermo@users.noreply.github.com>