gql to get the time that an asset's freshness state changed (#32337)
We currently compute this in the UI but that requires fetching the freshness policy definition from the asset graph and only works for time window policies. This adds a GQL endpoint that gets the time that the freshness record in the DB updated.



Also updates the `from_db_row`​ method of `FreshnessStateRecord`​ to use dot-access to get the entries we want from the row. There was a bug where the `updated_at`​ attr was sourced from the `created_at`​ column. Also adds timezone info to the datetime object since we store a utc datetime in the DB but read it as a datetime with no timezone. this was causing conversion issues when getting the timestamp to send to the front end



For future perf improvements we can move this to streamline https://linear.app/dagster-labs/issue/OPER-2443/source-timestamps-used-in-the-homepage-from-streamline