[ui] Unify timing and pez for jobs/assets (#32368)
## Summary & Motivation

The right-side elements of the job list and catalog list are similar but different, including the "pez" list that shows recent runs and materializations.

Unify these a bit better.

Asset catalog:

<img width="405" height="265" alt="Screenshot 2025-10-01 at 15 33 29" src="https://github.com/user-attachments/assets/7e382c64-9745-4bba-a6d9-745cbe160076" />

Jobs:

<img width="350" height="256" alt="Screenshot 2025-10-01 at 14 52 15" src="https://github.com/user-attachments/assets/a7682875-81a7-4144-9713-cd9b05ade78e" />


## How I Tested These Changes

View Catalog and Job list. Verify that the pez and timing information renders roughly the same in both.