add context hooks for definition-level permissions (#32151)
## Summary & Motivation
To enable definition-level permissions, we have to add some new concepts
to the request context and shift around some logic to start extracting
viewer/owner concepts from individual definitions.

This is a proposal for how to do that, specifically for the asset graph.
We'll have to extend the permission checks accordingly for
jobs/sensors/schedules (which might be tricky based on snapshots etc).

## How I Tested These Changes
BK