[ui] Add TIME_SECONDS to ReportingUnitType (#32387)
## Summary & Motivation

Corresponds to https://github.com/dagster-io/internal/pull/18482.

Add support for `TIME_SECONDS` as a reporting unit type.

`ReportingUnitType` is unfortunately defined in the internal GraphQL schema but used in OSS code, so we end up having to mirror the enum.

Added some tests for `formatMetric`, which are pretty basic but worth adding anyway.

## How I Tested These Changes

BK. Verify that `TIME_SECONDS`\-based values are rendered with the correct formatting in Insights.