Speed up partitions def calculation when a small subset of the asset graph is being computed (#32402)
## Summary & Motivation
Launchpad is passing in a list of asset keys to a full job. Before, we enumerated every asset in the asset job, fetched all the nodes, then intersected. Now, we intersect first before fetching nodes.

## How I Tested These Changes
Dry run on a query that was formerly fetching the whole list of nodes now only fetches one