Make "View Error" link on backfill errors no longer invisible
Summary:
This currently renders as a white link that is almost completely invisible against the salmon background of the modal. It is similarly almost invisible in dark mode. From what I can tell it was white way back in https://github.com/dagster-io/dagster/pull/11656 when it was first introduced so maybe it has never been visible? But now it is.

Launch a failed backfill in both light mode and dark mode.

[dagster-ui] Fixed an issue where the "View error" link in the popup that displayed when a backfill failed to launch was very difficult to see.

> Insert changelog entry or delete this section.