[ui] Prefer the global graph with a group filter over asset group pages
We’d like to remove the asset group pages which live at paths scoped to a code location.  We can’t /quite/ get rid of these pages yet, but this PR reduces the number of places in the app that link to them.

Instead of linking to the asset group page, we link to the global graph with a `group: “groupname”` filter.

After this PR, the groups pages are really only accessible via the old Jobs left nav.