FE-776: Disable the Materialize button if it’s scope is stale
This is a tiny fix for https://linear.app/dagster-labs/issue/FE-776/lineage-view-materialize-action-not-reflecting-asset-selection-if

If you type a query into the asset graph filter box both the graph data and layout can change. A loading bar appears, but while the UI is loading, clicking the Materialize button does not reflect your new filters.