[ui] Prefer the global graph with a group filter over asset group pages (#32407)
## Summary & Motivation

Related:
https://linear.app/dagster-labs/issue/FE-982/make-lineage-ignore-code-locations-in-groupings

We’d like to remove the asset group pages which live at paths scoped to
a code location. We can’t /quite/ get rid of these pages yet, but this
PR reduces the number of places in the app that link to them.

Instead of linking to the asset group page, we link to the global graph
with a `group: “groupname”` filter. The impacted callsites are mostly 1)
"View in group" (menu option on assets), 2) links to the group name on
the definition page, etc., and 3) the root page of the app when assets
and no jobs are present.

After this PR, the groups pages are really only accessible via the old
Jobs left nav.

## How I Tested These Changes

I tested these manually and added a new helper.

---------

Co-authored-by: bengotow <bgotow@elementl.com>