Enable adding output_metadata via build_output_context (#32371)
## Summary & Motivation
While working on a PR for dagster-iceberg I found that
`build_output_context` doesn't expose the `output_metadata` attribute of
`OutputContext` for configuration.

I tried using `context.add_output_metadata`, but found that it
unintuitively does not actually mutate the `output_metadata` value of
the `OutputContext` instance, so it can't be used to add
`output_metadata` to an existing `OutputContext` instance created by
`build_output_context`.

These changes enable a user to pass `output_metadata` to
`build_output_context` and have it be set on the resulting
`OutputContext` object.

## How I Tested These Changes
Added a unit test covering the new functionality. Ran the rest of the
unit tests

## Changelog

- Enable passing `output_metadata` into `build_output_context`