[dagster-dbt] Fix bug with DbtProjectComponent in which user-provided template vars would not be accessible from the `cli_args` field. (#32440)
## Summary & Motivation

See title.

This is a bit of a tricky one because we can't really update the signature of `execute()` (it's user-visible and intended to be overridden by subclasses), so we need to find some other way of making the original resolution scope available at runtime.

There are basically two options:

- eagerly render template values beforehand, and in the case that we hit an error with a variable not existing, just leave it untouched
- global state (chosen)

The first option seemed a little too funky to me (hard to conceptualize that there are two separate passes of rendering), so went with a global variable.

## How I Tested These Changes

## Changelog

[dagster-dbt] Fixed issue that made custom template vars unavailable when specifying them for the `cli_args:` field of the `DbtProjectComponent`.