[ui] Handle asset trend pez for >5 events (#32453)
## Summary & Motivation

When providing more than 5 events to `AssetRecentUpdatesTrend`, we end
up with an invalid array length on the array constructed for empty
items.

Resolve this by:

- Only using five events (or four, if latest info is provided) from the
provided `events` array
- Use a `Math.max` to guard against negative array lengths. This
shouldn't be strictly necessary now, since the `events` prop is being
sliced, but it shouldn't hurt.

Added a test as well.

## How I Tested These Changes

`yarn jest AssetRecentUpdatesTrend`

Load an asset catalog list that is known to show errors, verify that it
no longer does.