Use a different key for the RemoteExecutionPlan loader
Summary:
We didn't end up actually needing this in the change that I originally added it for, but I think it is still a good change for the future, since you don't need the run config to load a RemoteJob object. This in theory unblocks loading multiple execution plans for the same job with different run configs without having to refetch the same subset multiple times.

Test Plan: BK