replace streaming grpc repository gets with regular gets (#32484)
## Summary & Motivation


These were originally made streaming in a misguided attempt to work around byte size limits. There is no reason for the call to be streaming and some circumstantial evidence that it being streaming may be causing problems. This PR makes it synchronous, making sure that the higher timeout is kept.

## How I Tested These Changes

BK