[sbc] Store state in reconstruction metadata optionally (#32413)
## Summary & Motivation

Adds a third mode for state backed components which uses the "traditional" reconstruction metadata approach.

This approach has the following downsides (hence the revamped system):

- It refreshes every time the code server is reloaded, meaning there is no stability and no way to roll back.
- The above also means that if there's some transient issue with hitting the remote system, this will prevent the code location from loading entirely
- Finally, storing large amounts of serialized metadata in execution plan snapshots is not the most scalable thing in the world compared to the other approaches

However, it's still the easiest / simplest thing on the user's end as it requires no special infra or setup, and we have many pre-existing integrations that use this method. So having this integrate transparently with the rest of the StateBackedComponent infra and allows users to opt in to the fancier bits with just a bit of config

## How I Tested These Changes

## Changelog

NOCHANGELOG