Fix bug with LazyDefinitions and code server lifecycle (#32489)
## Summary & Motivation

Fixes a fairly nasty bug that occurs when you use the `@definitions` (`LazyDefinitions`) decorator in a context where `StateBackedDefinitionsLoaders` (or some new `StateBackedComponent` functionality) are present.

The core issue here is that the intended lifecycle for `StateBackedDefinitionsLoader` is:

1. Code server begins to spin up, `DefinitionsLoadContext` is in the `INITIALIZATION` state.
2. We load the definitions object as part of this process.
3. The `StateBackedDefinitionsLoaders` see that they are in the `INITIALIZATION` state, and so do their work to fetch the external state and write it to the context object.
4. Now that the RemoteRepoistory is loaded, we tack on that accumulated state.
5. We reconstruct the repository def from a pointer that includes that accumulated state.
6. We load the definitions again, but this time they don't have to fetch.

When you use a LazyDefinitions object, this gets messed up, because step (2) in the above process doesn't end up actually invoking any of the `StateBackedDefinitionsLoaders`, and so no state is accumulated during this process. The net result is that we end up needing to reftech the state whenever we build definitions, which is obviously bad.

This is a pretty slim fix that just makes sure that the code server actually invokes the lazy definitions function while it's still in `INITIALIZATION` mode

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.