Don't include ignored nodes when generating run config for an implicit asset job (#32463)
Summary:
This feature was intended to ensure that if you subset a job you can still pass in the original run config and have the unrelated fields still work. That logic does not apply to implicit asset job and just bloats the schema with information about every asset.

This required changing the launching backfills with config logic to only include the config for the relevant ops.