Don't include unrelated resource keys in run config schema for asset selections (#32465)
Summary:
Similar to https://github.com/dagster-io/dagster/pull/32463, but for resources. One thing I need to confirm is that required_resource_keys here includes every resource that *can* be included (i.e. some op might use), not everything that *must* be included in the actual config dictionary, despite the "required" in the name.

As with ops, the resources field becomes Permissive rather than Shape in this case to minimize the likelihood of this being a breaking change while reducing schema bloat considerably.