[da] Fix issue with the `execution_failed` automation condition (#32479)
## Summary & Motivation

Previously, this would be true for an unpartitioned asset if the latest run for that asset failed, even if the asset successfully materialized within that run. This is different behavior from the partitioned asset implementation.

## How I Tested These Changes

unit

## Changelog

Fixed a bug with `AutomationCondition.execution_failed` that would cause it to be evaluated as `True` for an unpartitioned asset in cases where the latest run failed, but the asset itself materialized successfully before that failure.