[ui] Combine groups on the asset graph across code locations (#28502)
Reviving this old PR for
https://linear.app/dagster-labs/issue/FE-982/make-lineage-ignore-code-locations-in-groupings!

This PR changes the **default behavior** of the asset graph so that
assets in groups with the same name in separate code locations are
grouped into a single box. It also updates the asset graph sidebar to
remove the top level of folders representing code locations, since the
graph doesn't render a code location hierarchy.

You can revert to the old behavior via a new feature flag.

---------

Co-authored-by: bengotow <bgotow@elementl.com>