[ui] Move executionPlan fetch out of RunActionsMenuRunFragment (#32514)
## Summary & Motivation

It appears that `executionPlan` can be a prohibitively expensive field
to query across a whole page of runs in the runs feed. Since we don't
need it for anything except what's in the actions menu, add it to the
menu's already-existing lazy query and take it out of the main page
query.

## How I Tested These Changes

Load Runs page in a scenario where it is known to be very slow.

`RunsFeedRootQuery` goes from ~32s in prod to ~1s.

## Changelog

[ui] Improve loading performance of Runs page.